Rename file + namespaces
